+++ components/checks/active/no_sql_injection.rb
@@ -39,7 +39,7 @@ class Arachni::Checks::NoSqlInjection < Arachni::Check::Base
             format:                    [Format::APPEND],
             regexp:                    error_patterns,
             ignore:                    ignore_patterns,
-            param_flip:                true,
+            fuzz_names:                true,
             longest_word_optimization: true
         }
     end