+++ components/checks/active/sql_injection.rb
@@ -47,7 +47,7 @@ class Arachni::Checks::SqlInjection < Arachni::Check::Base
             format:                    [Format::APPEND],
             regexp:                    error_patterns,
             ignore:                    ignore_patterns,
-            param_flip:                true,
+            fuzz_names:                true,
             longest_word_optimization: true
         }
     end